docker/provisioner/schemas/requirements.txt
requests
botocore<1.13
boto3<1.10
cfnresponse
python-schema-registry-client==1.6.0
httpx>=0.14,<0.15
